Revert "Merge branch 'LevelEditor' into Sprint-4-Demo"
This reverts commit 2dfe5366db4acdcc3f1d2525fb4f40e96a72c819.

+++ Aenir Engine 3.0/src/AenirEngine/Graphics/Shader Files/CellShader/CellShaderPS.hlsl
@@ -8,8 +8,8 @@ cbuffer CB_CellShader {
 	float4 outerEdgeColor;
 	float4 glowColor;
 	float glowStr;
-	float innerEdgeThickness;
-	float outerEdgeThickness;
+	int innerEdgeThickness;
+	int outerEdgeThickness;
 	float specularPower;
 	float specularStrength;
 	float roughness;
